Move taproot keys to the keys module
We have a keys module, taproot keys should live in there.
